Add support for user-defined constraints on instructions.
Add support for user-defined constraints on instructions.
These are written as an optional braced-block at the end of the
instruction pattern. When present, they are evaluated when the
instruction is generated and should return true if the values
of the variable fields are OK, false if the generation should
be retried with another random number.

Note that there is not much sanity checking on what we are
being asked to evaluate!