Improve diagnostics if master and slave disagree on packet size
Improve risu's diagnostics if the master and slave disagree about the
size of the packets they are transmitting. This usually means that
the two sides disagreed about whether a load/store insn should undef,
so one side has reached a 'compare memory' request and the other is
doing 'compare registers'.