risugen: implement memory block setup for aarch64
Eventually we will need to split the script into an
arch-independent part and an arch-dependent part
like the rest of the code.

Big thing to note is that this changes the length of the memory
block for loads and stores to the actual documented value of
~2K bytes of random data. Before, ~2K 32-bit words of random data
were written, for a total of ~8K. Not sure if this was intended,
but seems a bug, which gets fixed here.

Signed-off-by: Claudio Fontana <claudio.fontana@linaro.org>