aarch64.risu: add LDNP / STNP 32bit/64bit
add ldnp and stnp.

Note that all loads and store suffer from the problem that
base register must be != target register.
This is in some cases a limitation of our framework
(in particular with stores),
in some cases it seems the Foundation models also do not
handle the case predictably (with loads).

Signed-off-by: Claudio Fontana <claudio.fontana@linaro.org>