risugen: add --not-pattern flag for convenience
The --pattern flag is useful for including whole classes of instructions.
You could also do specific instruction with it. However if you want a
whole class of instructions without a specific few instructions you'd
need to come up with a more complex regex.

With this you can for example to:

  risugen --pattern ".* A64" --not-pattern "ROR.*" aarch64.risu test3.bin